[refactor] 학습 트래커 컴포넌트 추출
