+++ src/pages/Main.module.css
@@ -5,48 +5,6 @@
     box-sizing: border-box;
 }
 
-.tracker {
-    height: 170px;
-    background-color: #ffffff;
-    margin-top: 75px;
-    margin-left: 46px;
-    margin-right: 46px;
-    border-radius: 10px;
-    padding: 18px 16px;
-    position: relative;
-}
-
-.year {
-    font-size: 17px;
-    font-weight: 500;
-    color: #d5d5d5;
-    position: absolute;
-    bottom: 17px;
-}
-
-.trackerGrid {
-    width: calc(100% - 93px);
-    height: calc(100% - 36px);
-    display: grid;
-    grid-template-rows: repeat(7, 20px);
-    grid-auto-flow: column;
-    gap: 4px;
-    position: absolute;
-    right: 16px;
-}
-
-.trackerCell {
-    width: 18px;
-    height: 18px;
-    border-radius: 3px;
-    background-color: #ebedf0;
-}
-
-.level1 { background-color: #c6e48b; }
-.level2 { background-color: #7bc96f; }
-.level3 { background-color: #239a3b; }
-.level4 { background-color: #196127; }
-
 .postContainer {
     width: 100%;
     height: 170px;
